producer/requirements.txt
kafka-python>=2.0
requests
pandas
dotenv